unit test for updating temp script files, only replace temp script files when upload timestamp is later than file modification date

+++ bin/scripts.py
@@ -1,13 +1,13 @@
 import path
+import time
 import os
 import base64
 import config
-import calendar
-import datetime
 import lockfile
 
 scheme = 'script'
 extension = 'r'
+uploaded_stanza_key = 'uploaded'
 
 
 def get_custom_scripts_path():
@@ -27,10 +27,22 @@ def create_files(service):
             filename = name + os.path.extsep + extension
             full_path = os.path.join(custom_scripts_path, filename)
 
+            # in case there is no uploaded attribute, add it with the current timestamp
+            if not hasattr(stanza, uploaded_stanza_key):
+                stanza = config.create_stanza(service, scheme, name, {
+                    'content': stanza.__getattr__('content'),
+                    uploaded_stanza_key: int(time.time())
+                })
+
             if os.path.exists(full_path):
-                os.remove(full_path)
-                create_file = True
-                # TODO: compare file creation date with stanza creation date
+                # compare file creation date with stanza creation date
+                modify_time = int(os.path.getmtime(full_path))
+                uploaded_time = int(getattr(stanza, uploaded_stanza_key))
+                if uploaded_time > modify_time:
+                    os.remove(full_path)
+                    create_file = True
+                else:
+                    create_file = False
             else:
                 create_file = True
 
@@ -66,7 +78,7 @@ def get(service, script_name):
 def add(service, name, content):
     config.create_stanza(service, scheme, name, {
         'content': base64.encodestring(content).replace('\n', ''),
-        'uploaded': calendar.timegm(datetime.datetime.now().utctimetuple())
+        uploaded_stanza_key: int(time.time())
     })
